refactor Storage Package class
